requirements.txt
Django>=4.2 djangorestframework>=3.14 
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9 
celery>=5.3 
redis>=5.0 
django-celery-beat>=2.5
django-cors-headers>=4.9.0
gunicorn>=21.2
dotenv==0.9.9